moves nested components to the root of the parent
