+++ src/signals/incident/components/form/ContainerSelect/Selector/LegendPanel/LegendPanel.tsx
