+++ src/signals/incident/components/form/ContainerSelect/Selector/LegendToggleButton/LegendToggleButton.tsx
