+++ src/signals/incident/components/form/ContainerSelect/Selector/LegendToggleButton/__tests__/LegendToggleButton.test.tsx
