+++ src/signals/incident/components/form/ContainerSelect/Selector/LegendToggleButton/index.tsx
