+++ src/signals/incident/components/form/ContainerSelect/Summary/Summary.js
